balthapp/requirements.txt
amqp==5.1.1
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
commonmark==0.9.1
Deprecated==1.2.13
Flask==2.2.2
Flask-Limiter==3.1.0
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
limits==3.2.0
MarkupSafe==2.1.2
mimetypes-magic==0.4.30
ordered-set==4.1.0
packaging==23.0
prompt-toolkit==3.0.36
Pygments==2.14.0
pytz==2022.7.1
rich==12.6.0
six==1.16.0
SQLAlchemy==2.0.0
typing_extensions==4.4.0
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.2.2
wrapt==1.14.1
